requirements.txt
pyyaml
odooly
mysql-connector-python
